requirements.txt
anthropic>=0.25.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
flask>=3.0.0
requests>=2.31.0
tweepy>=4.14.0
praw>=7.7.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
gunicorn>=21.2.0
